requirements.txt
# Tax Document Processor - Python Dependencies

# OCR and PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
pytesseract>=0.5.10
pdf2image>=1.16.0
Pillow>=10.0.0

# LLM Integration
ollama>=0.1.0

# Vector Database
qdrant-client>=1.7.0

# Embeddings (for Qdrant)
sentence-transformers>=2.2.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Screen Capture
mss>=9.0.0
pyautogui>=0.9.54
keyboard>=0.13.5

# CLI Interface
click>=8.1.0
rich>=13.0.0

# Data Validation
pydantic>=2.0.0

# Database
# sqlite3 is built-in, no pip install needed

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0